services/analytics/requirements.txt
redis==5.0.7
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
vaderSentiment==3.3.2
langdetect==1.0.9
